+++ contracts/ownership/HasNoTokens.sol
@@ -15,10 +15,10 @@ contract HasNoTokens is Ownable {
  /** 
   * @dev Reject all ERC23 compatible tokens
   * @param from_ address The address that is transferring the tokens
-  * @param value_ Uint the amount of the specified token
+  * @param value_ uint256 the amount of the specified token
   * @param data_ Bytes The data passed from the caller.
   */
-  function tokenFallback(address from_, uint value_, bytes data_) external {
+  function tokenFallback(address from_, uint256 value_, bytes data_) external {
     throw;
   }
