+++ contracts/token/CrowdsaleToken.sol
@@ -15,13 +15,13 @@ contract CrowdsaleToken is StandardToken {
 
   string public constant name = "CrowdsaleToken";
   string public constant symbol = "CRW";
-  uint public constant decimals = 18;
+  uint256 public constant decimals = 18;
   // replace with your fund collection multisig address
   address public constant multisig = 0x0;
 
 
   // 1 ether = 500 example tokens
-  uint public constant PRICE = 500;
+  uint256 public constant PRICE = 500;
 
   /**
    * @dev Fallback function which receives ether and sends the appropriate number of tokens to the 
@@ -40,7 +40,7 @@ contract CrowdsaleToken is StandardToken {
       throw;
     }
 
-    uint tokens = msg.value.mul(getPrice());
+    uint256 tokens = msg.value.mul(getPrice());
     totalSupply = totalSupply.add(tokens);
 
     balances[recipient] = balances[recipient].add(tokens);
@@ -54,7 +54,7 @@ contract CrowdsaleToken is StandardToken {
    * @dev replace this with any other price function
    * @return The price per unit of token. 
    */
-  function getPrice() constant returns (uint result) {
+  function getPrice() constant returns (uint256 result) {
     return PRICE;
   }
 }
