+++ test/helpers/VestedTokenMock.sol
@@ -4,7 +4,7 @@ import '../../contracts/token/VestedToken.sol';
 
 // mock class using StandardToken
 contract VestedTokenMock is VestedToken {
-  function VestedTokenMock(address initialAccount, uint initialBalance) {
+  function VestedTokenMock(address initialAccount, uint256 initialBalance) {
     balances[initialAccount] = initialBalance;
     totalSupply = initialBalance;
   }
